+++ R/write_bds.R
@@ -22,7 +22,7 @@
 #' @seealso [jsonlite::toJSON()]
 #' @examples
 #' fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
-#' tgt <- read_bds(fn, append_ddi = TRUE, schema = "bds_schema_str.json")
+#' tgt <- read_bds(fn, append_ddi = TRUE, schema = "bds_schema_v1.0.json")
 #' js <- write_bds(tgt)
 #' @export
 write_bds <- function(x = NULL, file = NULL, schema = NULL, indent = NULL,