+++ NEWS.md
@@ -14,7 +14,7 @@
 
 * Set `bds_schema.json` as default!! (not all testfile yet work)
 * Replaces the `jamestest` package by `jamesdemodata`
-* Supports (testing of) `bds_schema_str.json` (old) and `bds_schema.json` (new)
+* Supports (testing of) `bds_schema_v1.0.json` (old) and `bds_schema.json` (new)
 * Adds a `file` argument to `write_bds()`
 * Updates scripts to account for double schema